** UI: Implement (H/V)ProgressBar & (H/V)ScrollBar MainUI Rearrangement * CORE: Add II18NConvertor. Adjust LpsAudioCore           & MusicEntity           & IMeidaMetaDataReader

+++ Lunalipse.Common/Interfaces/IAudio/ILpsAudio.cs
@@ -16,6 +16,6 @@ namespace Lunalipse.Common.Interfaces.IAudio
         void Pause();
         void Resume();
         void Stop();
-        void MoveTo(long secs);
+        void MoveTo(double secs);
     }
 }
